htsget local ingest + refactoring
