+++ README.md
@@ -80,17 +80,39 @@ python s3_ingest.py --sample <sample>|--samplefile <samplefile> --endpoint <S3 e
 
 </details></blockquote>
 
+### Store locally in htsget
+If necessary, genomic samples can be loaded directly from the htsget container's internal storage. Simply `docker cp` the sample files somewhere into the container and proceed below with the instructions for local ingest.
+
 ### Ingest into Htsget
 
-Create a text file that list all sample IDs available in a particular S3 bucket. The ingest script will find all files starting with that particular ID in that bucket; for example, specifying AB0001 will ingest, if available, AB0001.vcf.gz/tbi, AB0001.mutect2.vcf.gz/tbi, and AB0001_comparison.vcf.gz/tbi. The bucket should contain both bgzipped VCF files and their corresponding tabix index files.
+Genomic samples should be specified in a JSON list of dictionaries, providing their genomic ID and optionally associated clinical IDs and filenames.
+The respective keys for these attributes are genomic_id, clinical_id, and files; if filenames are specified, the files key
+should be another dictionary with the keys "index" and "sample", which specify the genomic index file and variation file respectively.
+For example:
 
-Connecting the genomic IDs and files to patient clinical data will be handled during clinical data ingest; see below.
-
-To make the genomic files accessible to the htsget server, run:
-
-```bash
-python htsget_s3_ingest.py --sample <sample>|--samplefile <samplefile> --dataset <dataset>  --awsfile <aws credentials> --endpoint <S3 endpoint> --bucket <S3 bucket> --prefix <optional, prefix for files in S3 bucket> --reference <reference genome, either hg37 or hg38>
+```json
+[
+  {"genomic_id": "HG00096", "clinical_id": null, "files": null}, 
+  {"genomic_id": "HG00097", "clinical_id": "DONOR_1", "files": {"sample": "HG97_SAMPLE.vcf.gz", "index": "HG97_SAMPLE.vcf.gz.tbi"}}
+  {"genomic_id": "HG00099", "clinical_id": null, "files": null}
+]
 ```
+If files are stored in an S3 Bucket, the location of the files should be provided using the prefix argument in htsget_ingest, and only filenames should be given in the JSON.
+For local ingest, absolute paths within the container should be provided for filenames.
+
+If filenames are not provided, the ingest will search for any variation/index files in the container prefixed with the genomic_id.
+For local ingest, filenames must be specified.
+
+To ingest using an S3 container, once the files have been added, you can run the htsget_ingest.py script:
+```bash
+python htsget_ingest.py --samplefile [JSON-formatted sample list as specified] --dataset <dataset>  --awsfile <aws credentials> --endpoint <S3 endpoint> --bucket <S3 bucket> --prefix <optional, prefix for files in S3 bucket> --reference <reference genome, either hg37 or hg38> --indexing <optional, force re-index>
+```
+
+To ingest from the htsget container, you can use the -local flag:
+```bash
+python htsget_ingest.py -local --samplefile [JSON-formatted sample list] --dataset <dataset>  --reference <reference genome, either hg37 or hg38> --indexing <optional, force re-index>
+```
+
 
 ## Ingest clinical data
 
@@ -146,10 +168,11 @@ Genomic data can be ingested from an S3 bucket at the /ingest_genomic endpoint, 
     "bucket": "[S3 bucket name]",
     "access": "[S3 bucket access username]",
     "secret": "[S3 bucket access password]",
-    "samples": ["[sample name 1]", "[sample name 2]", ...],
+    "samples": [JSON formatted samples, see command line instructions],
     "prefix": "[S3 prefix, optional]",
     "reference": "[Reference genome, either hg37 or hg38, optional]",
-    "indexing": "[Force reindexing (true/false), optional]"
+    "indexing": "[Force reindexing (true/false), optional]",
+    "local": true/false (whether to ingest from htsget local storage instead of s3)
 }
 ```
 Make sure you include Authorization headers as well. Both endpoints take refresh tokens, as a header with the key `{"Authorization": "Bearer [refresh token]"}`.
